+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.0.0](https://github.com/in2workspace/in2-verifier-api/releases/tag/v2.0.0)
+- New major version to align with the new major version of EUDIStack project.
+
+## [v1.3.11](https://github.com/in2workspace/in2-verifier-api/releases/tag/v1.3.10)
+### Added
+- Added revocation function for new credentials with credentialStatus.
+- Test for verify that is working the revocation
+
 ## [v1.3.10](https://github.com/in2workspace/in2-verifier-api/releases/tag/v1.3.10)
 ### Added
 - Added access for prometheus at spring security at matcher.